upcoming: [DI-20800] - Rename files
