+++ packages/manager/src/features/CloudPulse/Widget/components/Zoomer.tsx
@@ -1,8 +1,8 @@
 import { useTheme } from '@mui/material';
 import * as React from 'react';
 
-import ZoomInMap from 'src/assets/icons/cloudpulse_zoomin.svg';
-import ZoomOutMap from 'src/assets/icons/cloudpulse_zoomout.svg';
+import ZoomInMap from 'src/assets/icons/zoomin.svg';
+import ZoomOutMap from 'src/assets/icons/zoomout.svg';
 
 export interface ZoomIconProperties {
   className?: string;
