+++ packages/manager/src/features/CloudPulse/shared/CloudPulseDashboardFilterBuilder.tsx
@@ -1,8 +1,8 @@
 import { Grid, Typography } from '@mui/material';
 import * as React from 'react';
 
-import KeyboardArrowDownIcon from 'src/assets/icons/cloudpulse_arrow_down.svg';
-import KeyboardArrowRightIcon from 'src/assets/icons/cloudpulse_arrow_right.svg';
+import KeyboardArrowDownIcon from 'src/assets/icons/arrow_down.svg';
+import KeyboardArrowRightIcon from 'src/assets/icons/arrow_right.svg';
 import InfoIcon from 'src/assets/icons/info.svg';
 import { Button } from 'src/components/Button/Button';
 import { ErrorState } from 'src/components/ErrorState/ErrorState';
